Create Strings in alphabetical order
Ria is a 5 year old girl. Her mother wants to teach her how to sort words in the same order that they appear in a dictionary. She decided to write a program to sort a given set of strings based on their alphabetical order. Help Ria's mother to complete the program.
INPUT FORMAT
A set of N strings
OUTPUT FORMAT
Alphabetically sorted set of strings